Refactor into Preauth and Postauth
Renamed folder preauth (for login/register and other functions for unauthenticated) and postauth (for all else).
